Fix all clippy warnings
- Suppress unused_assignments from miette Diagnostic derive macro
- Remove unnecessary i64 cast in lowering
- Use slice instead of &mut Vec in fold_constants
- Collapse nested if-let expressions
- Derive Default for Resolver
- Remove unused enumerate index

+++ src/bir/lowering.rs
@@ -320,7 +320,7 @@ impl Lowering {
                 let result = self.fresh_value();
                 self.emit(Instruction::Literal {
                     result,
-                    value: *n as i64,
+                    value: *n,
                     ty: BirType::I32,
                 });
                 self.value_types.insert(result, BirType::I32);
